index/requirements.txt
Faker==17.6.0
psycopg[binary]
python-dotenv==1.0.0
